Move inf-server stuff to new module
